IAE is too broad and the class not found requires special handling. ClassNotFoundRuntimeException is more appropriate for this error.
