+++ src/main/java/net/openhft/chronicle/wire/TextWire.java
@@ -1878,7 +1878,7 @@ public class TextWire extends YamlWireOut<TextWire> {
                         }
                         String message = "Unable to load " + stringBuilder + ", is a class alias missing.";
                         if (IAE_ON_CNF)
-                            throw new IllegalArgumentException(message);
+                            throw new ClassNotFoundRuntimeException(new ClassNotFoundException(message));
                         Jvm.warn().on(TextWire.class, message);
                         return null;
                     }
@@ -1899,7 +1899,7 @@ public class TextWire extends YamlWireOut<TextWire> {
                         }
 
                     } else if (tClass.getClassLoader() == null) {
-                        throw new IllegalArgumentException("Unable to find class " + stringBuilder);
+                        throw new ClassNotFoundRuntimeException(new ClassNotFoundException("Unable to find class " + stringBuilder));
                     } else {
                         return Wires.tupleFor(tClass, stringBuilder.toString());
                     }
