+++ src/test/java/net/openhft/chronicle/wire/security/ApacheStructExploitTest.java
@@ -20,6 +20,7 @@ package net.openhft.chronicle.wire.security;
 
 
 import net.openhft.chronicle.core.Jvm;
+import net.openhft.chronicle.core.util.ClassNotFoundRuntimeException;
 import net.openhft.chronicle.wire.Marshallable;
 import net.openhft.chronicle.wire.WireTestCommon;
 import org.junit.Test;
@@ -57,7 +58,7 @@ public class ApacheStructExploitTest extends WireTestCommon {
                 fail();
             assertEquals("{dataHandler={dataSource={contentType=null, is=javax.crypto.CipherInputStream@xxxx, consumed=false}, objDataSource=null, object=null, objectMimeType=null, currentCommandMap=null, transferFlavors=[], dataContentHandler=null, factoryDCH=null, oldFactory=null, shortType=null}, data=null, dataLen=0, mimeType=null}",
                     fromString.toString().replaceFirst("@[0-9a-f]+", "@xxxx"));
-        } catch (IllegalArgumentException iae) {
+        } catch (ClassNotFoundRuntimeException iae) {
             assertTrue(Jvm.isJava9Plus());
 
         } catch (IllegalStateException ise) {
